Refactoring to reduce the amount of components for consistent page spacing and transition
